added AddressMap buildera
