+++ go/store/prolly/map.go
@@ -20,7 +20,11 @@ import (
 	"io"
 	"strings"
 
+	fb "github.com/google/flatbuffers/go"
+
+	"github.com/dolthub/dolt/go/gen/fb/serial"
 	"github.com/dolthub/dolt/go/store/hash"
+	"github.com/dolthub/dolt/go/store/pool"
 	"github.com/dolthub/dolt/go/store/prolly/tree"
 	"github.com/dolthub/dolt/go/store/types"
 	"github.com/dolthub/dolt/go/store/val"
@@ -50,7 +54,7 @@ func NewMap(node tree.Node, ns tree.NodeStore, keyDesc, valDesc val.TupleDesc) M
 
 // NewMapFromTuples creates a prolly tree Map from slice of sorted Tuples.
 func NewMapFromTuples(ctx context.Context, ns tree.NodeStore, keyDesc, valDesc val.TupleDesc, tups ...val.Tuple) (Map, error) {
-	ch, err := tree.NewEmptyChunker(ctx, ns, newNodeBuilder)
+	ch, err := tree.NewEmptyChunker(ctx, ns, newMapBuilder)
 	if err != nil {
 		return Map{}, err
 	}
@@ -239,6 +243,103 @@ func (p *pointLookup) Next(context.Context) (key, value val.Tuple, err error) {
 		p.k, p.v = nil, nil
 	}
 	return
+}
+
+func newEmptyMapNode(pool pool.BuffPool) tree.Node {
+	bld := &mapBuilder{}
+	return bld.Build(pool)
+}
+
+func newMapBuilder(level int) *mapBuilder {
+	return &mapBuilder{level: level}
+}
+
+var _ tree.NodeBuilderFactory[*mapBuilder] = newMapBuilder
+
+type mapBuilder struct {
+	keys, values []tree.Item
+	size, level  int
+
+	subtrees tree.SubtreeCounts
+}
+
+var _ tree.NodeBuilder = &mapBuilder{}
+
+func (nb *mapBuilder) StartNode() {
+	nb.reset()
+}
+
+func (nb *mapBuilder) HasCapacity(key, value tree.Item) bool {
+	sum := nb.size + len(key) + len(value)
+	return sum <= int(tree.MaxVectorOffset)
+}
+
+func (nb *mapBuilder) AddItems(key, value tree.Item, subtree uint64) {
+	nb.keys = append(nb.keys, key)
+	nb.values = append(nb.values, value)
+	nb.size += len(key) + len(value)
+	nb.subtrees = append(nb.subtrees, subtree)
+}
+
+func (nb *mapBuilder) Count() int {
+	return len(nb.keys)
+}
+
+func (nb *mapBuilder) reset() {
+	// buffers are copied, it's safe to re-use the memory.
+	nb.keys = nb.keys[:0]
+	nb.values = nb.values[:0]
+	nb.size = 0
+	nb.subtrees = nb.subtrees[:0]
+}
+
+func (nb *mapBuilder) Build(pool pool.BuffPool) (node tree.Node) {
+	var (
+		keyTups, keyOffs fb.UOffsetT
+		valTups, valOffs fb.UOffsetT
+		refArr, cardArr  fb.UOffsetT
+	)
+
+	keySz, valSz, bufSz := estimateBufferSize(nb.keys, nb.values, nb.subtrees)
+	b := getFlatbufferBuilder(pool, bufSz)
+
+	// serialize keys and offsets
+	keyTups = writeItemBytes(b, nb.keys, keySz)
+	serial.ProllyTreeNodeStartKeyOffsetsVector(b, len(nb.keys)-1)
+	keyOffs = writeItemOffsets(b, nb.keys, keySz)
+
+	if nb.level == 0 {
+		// serialize value tuples for leaf nodes
+		valTups = writeItemBytes(b, nb.values, valSz)
+		serial.ProllyTreeNodeStartValueOffsetsVector(b, len(nb.values)-1)
+		valOffs = writeItemOffsets(b, nb.values, valSz)
+	} else {
+		// serialize child refs and subtree counts for internal nodes
+		refArr = writeItemBytes(b, nb.values, valSz)
+		cardArr = writeCountArray(b, nb.subtrees)
+	}
+
+	// populate the node's vtable
+	serial.ProllyTreeNodeStart(b)
+	serial.ProllyTreeNodeAddKeyItems(b, keyTups)
+	serial.ProllyTreeNodeAddKeyOffsets(b, keyOffs)
+	if nb.level == 0 {
+		serial.ProllyTreeNodeAddValueItems(b, valTups)
+		serial.ProllyTreeNodeAddValueOffsets(b, valOffs)
+		serial.ProllyTreeNodeAddTreeCount(b, uint64(len(nb.keys)))
+	} else {
+		serial.ProllyTreeNodeAddAddressArray(b, refArr)
+		serial.ProllyTreeNodeAddSubtreeCounts(b, cardArr)
+		serial.ProllyTreeNodeAddTreeCount(b, nb.subtrees.Sum())
+	}
+	serial.ProllyTreeNodeAddKeyType(b, serial.ItemTypeTupleFormatAlpha)
+	serial.ProllyTreeNodeAddValueType(b, serial.ItemTypeTupleFormatAlpha)
+	serial.ProllyTreeNodeAddTreeLevel(b, uint8(nb.level))
+	b.Finish(serial.ProllyTreeNodeEnd(b))
+	nb.reset()
+
+	buf := b.FinishedBytes()
+	return tree.NodeFromBytes(buf)
 }
 
 // DebugFormat formats a Map.
