+++ go/store/prolly/serialize.go
@@ -1,0 +1,89 @@
+// Copyright 2021 Dolthub, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package prolly
+
+import (
+	"encoding/binary"
+	"fmt"
+
+	fb "github.com/google/flatbuffers/go"
+
+	"github.com/dolthub/dolt/go/store/pool"
+	"github.com/dolthub/dolt/go/store/prolly/tree"
+)
+
+func getFlatbufferBuilder(pool pool.BuffPool, sz int) (b *fb.Builder) {
+	b = fb.NewBuilder(0)
+	buf := pool.Get(uint64(sz))
+	b.Bytes = buf[:0]
+	return
+}
+
+func writeItemBytes(b *fb.Builder, items []tree.Item, sumSz int) fb.UOffsetT {
+	b.Prep(fb.SizeUOffsetT, sumSz)
+
+	stop := int(b.Head())
+	start := stop - sumSz
+	for _, item := range items {
+		copy(b.Bytes[start:stop], item)
+		start += len(item)
+	}
+
+	start = stop - sumSz
+	return b.CreateByteVector(b.Bytes[start:stop])
+}
+
+func writeItemOffsets(b *fb.Builder, items []tree.Item, sumSz int) fb.UOffsetT {
+	var cnt int
+	var off = sumSz
+	for i := len(items) - 1; i > 0; i-- { // omit first offset
+		off -= len(items[i])
+		b.PrependUint16(uint16(off))
+		cnt++
+	}
+	return b.EndVector(cnt)
+}
+
+func writeCountArray(b *fb.Builder, sc tree.SubtreeCounts) fb.UOffsetT {
+	// todo(andy) write without copy
+	arr := tree.WriteSubtreeCounts(sc)
+	return b.CreateByteVector(arr)
+}
+
+// estimateBufferSize returns the exact Size of the tuple vectors for keys and values,
+// and an estimate of the overall Size of the final flatbuffer.
+func estimateBufferSize(keys, values []tree.Item, subtrees []uint64) (keySz, valSz, bufSz int) {
+	for i := range keys {
+		keySz += len(keys[i])
+		valSz += len(values[i])
+	}
+	refCntSz := len(subtrees) * binary.MaxVarintLen64
+
+	// constraints enforced upstream
+	if keySz > int(tree.MaxVectorOffset) {
+		panic(fmt.Sprintf("key vector exceeds Size limit ( %d > %d )", keySz, tree.MaxVectorOffset))
+	}
+	if valSz > int(tree.MaxVectorOffset) {
+		panic(fmt.Sprintf("value vector exceeds Size limit ( %d > %d )", valSz, tree.MaxVectorOffset))
+	}
+
+	bufSz += keySz + valSz               // tuples
+	bufSz += refCntSz                    // subtree counts
+	bufSz += len(keys)*2 + len(values)*2 // offsets
+	bufSz += 8 + 1 + 1 + 1               // metadata
+	bufSz += 72                          // vtable (approx)
+
+	return
+}
